backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
rq==1.15.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
pydantic-settings==2.1.0
